修复父非子的spDic是int类型导致综合spDic失败;

+++ SMG_NothingIsAll/AIFoundation/AINet/AIPort/AIPort.h
@@ -66,8 +66,8 @@
  */
 @interface AISPStrong : NSObject <NSCoding>
 
-@property (assign,nonatomic) NSInteger sStrong;
-@property (assign,nonatomic) NSInteger pStrong;
+@property (assign,nonatomic) CGFloat sStrong;
+@property (assign,nonatomic) CGFloat pStrong;
 
 @end
